Filtering the GRG in place with LinearOperators
For large-scale filtering, it is best to filter the GRG using
the utilities provided in this repo. However, if you only need to
"leave on out" or do other small modifications, it is faster to
just ignore the mutations/samples during calculation.

LinearOperators now support a GRGOperatorFilter that lets you
specify which mutations/samples to keep.